refactor: move resolveProvider function to src/providers/index.ts - Move resolveProvider function and LoadApiProviderOptions interface from src/index.ts to src/providers/index.ts for better organization - Update imports in src/index.ts to use resolveProvider from providers module - Remove unused imports (ProviderOptions, invariant) from src/index.ts - Update test mocks to allow resolveProvider to work with mocked loadApiProvider functions - 6 out of 8 providerMap tests passing (core functionality working correctly)
