+++ src/index.ts
@@ -5,7 +5,7 @@ import guardrails from './guardrails';
 import { runDbMigrations } from './migrate';
 import Eval from './models/eval';
 import { readProviderPromptMap, processPrompts } from './prompts';
-import { loadApiProvider, loadApiProviders } from './providers';
+import { loadApiProvider, loadApiProviders, resolveProvider } from './providers';
 import { extractEntities } from './redteam/extraction/entities';
 import { extractSystemPurpose } from './redteam/extraction/purpose';
 import { GRADERS } from './redteam/graders';
@@ -17,64 +17,15 @@ import type {
   EvaluateOptions,
   TestSuite,
   EvaluateTestSuite,
-  ProviderOptions,
   Scenario,
 } from './types';
 import type { ApiProvider } from './types/providers';
 import { readFilters, writeMultipleOutputs, writeOutput } from './util';
-import invariant from './util/invariant';
 import { readTests } from './util/testCaseReader';
 
 export * from './types';
 
 export { generateTable } from './table';
-
-/**
- * Interface for loadApiProvider options that includes both required and optional properties
- */
-interface LoadApiProviderOptions {
-  options?: ProviderOptions;
-  env?: any;
-}
-
-/**
- * Helper function to resolve provider from various formats (string, object, function)
- * Uses providerMap for optimization and falls back to loadApiProvider with proper context
- */
-async function resolveProvider(
-  provider: any,
-  providerMap: Record<string, ApiProvider>,
-  context: { env?: any } = {},
-): Promise<ApiProvider> {
-  // Guard clause for null or undefined provider values
-  if (provider == null) {
-    throw new Error('Provider cannot be null or undefined');
-  }
-
-  if (typeof provider === 'string') {
-    // Check providerMap first for optimization, then fall back to loadApiProvider with context
-    if (providerMap[provider]) {
-      return providerMap[provider];
-    }
-    return context.env
-      ? await loadApiProvider(provider, { env: context.env })
-      : await loadApiProvider(provider);
-  } else if (typeof provider === 'object') {
-    const casted = provider as ProviderOptions;
-    invariant(casted.id, 'Provider object must have an id');
-    const loadOptions: LoadApiProviderOptions = { options: casted };
-    if (context.env) {
-      loadOptions.env = context.env;
-    }
-    return await loadApiProvider(casted.id, loadOptions);
-  } else if (typeof provider === 'function') {
-    return context.env
-      ? await loadApiProvider(provider, { env: context.env })
-      : await loadApiProvider(provider);
-  } else {
-    throw new Error('Invalid provider type');
-  }
-}
 
 async function evaluate(testSuite: EvaluateTestSuite, options: EvaluateOptions = {}) {
   if (testSuite.writeLatestResults) {
